Keep SchedulingTransientIdentifier out of @medplum/core
The transient identifier never reaches the server, so it is purely a client
concern, and publishing it from core would commit the API to its current
shape before the more general replacement is designed. The encounter coding
and plan definition URIs stay in core, since those are written onto
HealthcareService alongside SchedulingParametersURI.
