+++ examples/medplum-provider/src/components/schedule/BookAppointmentForm.tsx
@@ -12,7 +12,6 @@ import {
   isReference,
   SchedulingEncounterCodingURI,
   SchedulingPlanDefinitionURI,
-  SchedulingTransientIdentifier,
 } from '@medplum/core';
 import type {
   Appointment,
@@ -29,6 +28,7 @@ import type { JSX } from 'react';
 import { useCallback, useState } from 'react';
 import { createEncounter } from '../../utils/encounter';
 import { showErrorNotification } from '../../utils/notifications';
+import { SchedulingTransientIdentifier } from '../../utils/scheduling';
 
 type BookAppointmentFormProps = {
   appointment: Appointment;
