+++ packages/core/src/scheduling.ts
@@ -1,17 +1,8 @@
 // SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
 // SPDX-License-Identifier: Apache-2.0
-import type { CodeableConcept, HealthcareService, Identifier, Reference, Resource, Schedule } from '@medplum/fhirtypes';
-import { generateId } from './crypto';
+import type { CodeableConcept, HealthcareService, Reference, Schedule } from '@medplum/fhirtypes';
 import type { WithId } from './utils';
-import {
-  createReference,
-  getExtension,
-  getExtensionValue,
-  getIdentifier,
-  getReferenceString,
-  isDefined,
-  setIdentifier,
-} from './utils';
+import { createReference, getExtension, getExtensionValue, getReferenceString, isDefined } from './utils';
 
 export const SchedulingParametersURI = 'https://medplum.com/fhir/StructureDefinition/SchedulingParameters';
 export const SchedulingEncounterCodingURI = 'https://medplum.com/fhir/StructureDefinition/SchedulingEncounterCoding';
@@ -115,27 +106,3 @@ export function extractServiceTypeReferences(
     .map((concept) => getExtensionValue(concept, ServiceTypeReferenceURI) as Reference<HealthcareService> | undefined)
     .filter(isDefined);
 }
-
-const MedplumSchedulingTransientIdentifierURI = 'https://medplum.com/fhir/scheduling-transient-id';
-
-/**
- * Tags resources that do not exist on the server yet, such as the proposed
- * Appointments and Slots returned by `$find`, with a `use: 'temp'` identifier.
- * This gives clients a stable key for a resource that has no `id`. Call
- * `remove` before persisting, so the identifier never reaches the server.
- */
-export const SchedulingTransientIdentifier = {
-  set(resource: Resource & { identifier?: Identifier[] }) {
-    setIdentifier(resource, MedplumSchedulingTransientIdentifierURI, generateId(), { use: 'temp' });
-  },
-
-  get(resource: Resource) {
-    return getIdentifier(resource, MedplumSchedulingTransientIdentifierURI);
-  },
-
-  remove(resource: Resource & { identifier?: Identifier[] }) {
-    resource.identifier = resource.identifier?.filter(
-      (identifier) => identifier.system !== MedplumSchedulingTransientIdentifierURI
-    );
-  },
-};
